pyproject.toml
[tool.poetry]
name = "github-tools"
version = "0.3.2"
description = "Helper tools for GitHub"
authors = ["David Wales <david.wales@swsphn.com.au>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8"
requests = ">=2.31.0"
typer = ">=0.9.0"
azure-identity = ">=1.15.0"
azure-keyvault-keys = ">=4.8.0"
rich = "^13.7.0"

[tool.poetry.group.dev.dependencies]
black = ">=23.12.0"

[tool.poetry.scripts]
ght = 'github_tools:app'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
